Get test that is attached to object as "test" attribute

The returned function can be executed. If that is done within a testthat::Reporter, the results will be reported accordingly. Without reporter, the test runs silently if successful and stops if it fails.

get_test(object)

Arguments

object
to which the test is attached

Value

the test (function)

Examples

f <- function(x) x^2

test(f) <- function() {
   context("f")

   test_that("correct result for complex number", {
     expect_equal(f(1i), -1 + 0i)
   })
}

get_test(f)
#> function() {
#>    context("f")
#> 
#>    test_that("correct result for complex number", {
#>      expect_equal(f(1i), -1 + 0i)
#>    })
#> }
#> <environment: 0x7fa18f67bfb8>

library(testthat)
## execute the test
get_test(f)()
#> Test passed ğŸ¥‡

with_reporter("summary", get_test(f)())
#> f: .
#> â•â• DONE â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•